pyproject.toml
[project]
name = "py-track-visualization"
version = "0.1.0"
description = ""
authors = [
    {name = "Max",email = "mxbely98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dash (>=2.18.2,<3.0.0)",
    "dash-leaflet (>=1.0.15,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "ruff (>=0.9.8,<0.10.0)",
    "isort (>=6.0.1,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
